// rickandmorty-api/src/types/Characters.ts
export type Character = {
    id: number;
    name: string;
    status: string; 
    species: string;
    type: string;
    gender: string; 
    origin: Location; 
    location: Location; 
    image: string; 
    episode: string[];
    url: string; 
    created: string;
  };
  interface Location {
    name: string; 
    url: string; 
  }